Fix ClientController

+++ backend/app/Http/Controllers/ClientController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Client;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ClientController extends Controller
 {
@@ -12,9 +13,10 @@ class ClientController extends Controller
      */
     public function index()
     {
-        $clients = Client::all();
-        return response()->json($clients);
+        $authUser = Auth::user();
 
+        $orders = Client::where('company_id', $authUser->company_id)->get();
+        return response()->json($orders);
     }
 
 
@@ -23,6 +25,7 @@ class ClientController extends Controller
      */
     public function store(Request $request)
     {
+        $authUser = Auth::user();
         $validatedData = $request->validate([
             'last_name' => 'required|string|max:50',
             'first_name' => 'required|string|max:50',
@@ -34,11 +37,9 @@ class ClientController extends Controller
             'billing_city' => 'required|string|max:50',
             'email' => 'required|email:rfc,dns',
             'phone' => 'required|string|max:10',
-            'company_id' => 'required|numeric'
         ]);
-
+        $validatedData['company_id'] = $authUser->company_id;
         $client = Client::create($validatedData);
-
         return response()->json($client, 201);
     }
 
@@ -47,15 +48,18 @@ class ClientController extends Controller
      */
     public function show($id)
     {
-
+        $authUser = Auth::user();
         $client = Client::find($id);
 
         if (!$client) {
             return response()->json(['error' => 'Client not found'], 404);
         }
 
-        return response()->json($client);
+        if ($client->company_id !== $authUser->company_id) {
+            return response()->json(['error' => 'Forbidden'], 403);
+        }
 
+        return response()->json($client);
     }
 
     /**
@@ -63,10 +67,15 @@ class ClientController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $authUser = Auth::user();
         $client = Client::find($id);
 
         if (!$client) {
             return response()->json(['error' => 'Client not found'], 404);
+        }
+
+        if ($client->company_id !== $authUser->company_id) {
+            return response()->json(['error' => 'Forbidden'], 403);
         }
 
         $validatedData = $request->validate([
@@ -80,22 +89,27 @@ class ClientController extends Controller
             'billing_city' => 'required|string|max:50',
             'email' => 'required|email:rfc,dns',
             'phone' => 'required|string|max:10',
-            'company_id' => 'required|numeric'
         ]);
 
+        $validatedData['company_id'] =$authUser->company_id;
         $client->update($validatedData);
-
-        return response()->json($client);    }
+        return response()->json($client);    
+    }
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy($id)
     {
+        $authUser = Auth::user();
         $client = Client::find($id);
 
         if (!$client) {
             return response()->json(['error' => 'Client not found'], 404);
+        }
+
+        if ($client->company_id !== $authUser->company_id) {
+            return response()->json(['error' => 'Forbidden'], 403);
         }
 
         $client->delete();
